Procedimiento almacenados 1/2
Se realizo la mitad de los procedimiento almacenados a su vez se comprobó su buen funcionamiento.

+++ script/Script1 - Tablas.sql
@@ -302,7 +302,7 @@ CREATE TABLE Motivo(
 -- Creacion de la tabla Inflacion
 CREATE TABLE Inflacion(
 	anno INT,
-	porcentaje INT,
+	porcentaje float,
 	PRIMARY KEY(anno)
 );
